Aggregators test added

+++ pavellevap/src/main/java/ru/fizteh/fivt/students/pavellevap/CQL/aggregatesImpl/Avg.java
@@ -2,23 +2,23 @@ package ru.fizteh.fivt.students.pavellevap.CQL.aggregatesImpl;
 
 import java.util.function.Function;
 
-public class Avg<T> implements Aggregator<T, Long> {
+public class Avg<T> implements Aggregator<T, Double> {
     private Function<T, ? extends Number> function;
     public Avg(Function<T, ? extends Number> function) {
         this.function = function;
     }
 
-    public Long apply(Iterable<T> elements) {
-        Long result = 0L;
+    public Double apply(Iterable<T> elements) {
+        Double result = 0.0;
         Integer count = 0;
         for (T element : elements) {
-            result += (Long) function.apply(element);
+            result += function.apply(element).doubleValue();
             count++;
         }
         return result / count;
     }
 
-    public Long apply(T t) {
+    public Double apply(T t) {
         return null;
     }
 }
